+++ internal/db/queries/cards.sql
@@ -1,0 +1,2 @@
+-- name: GetCard :one
+SELECT * FROM cards WHERE id = $1;