+++ internal/db/queries/deck_access.sql
@@ -1,0 +1,2 @@
+-- name: GetDeckAccess :one
+SELECT * FROM deck_access WHERE deck_id = $1 AND user_id = $2;