+++ internal/db/queries/fields.sql
@@ -1,0 +1,2 @@
+-- name: GetField :one
+SELECT * FROM fields WHERE id = $1;