+++ internal/db/queries/media_blobs.sql
@@ -1,0 +1,2 @@
+-- name: GetMediaBlob :one
+SELECT * FROM media_blobs WHERE sha256 = $1;