+++ internal/db/queries/review_log.sql
@@ -1,0 +1,2 @@
+-- name: GetReviewLogEntry :one
+SELECT * FROM review_log WHERE id = $1;