+++ internal/db/queries/templates.sql
@@ -1,0 +1,2 @@
+-- name: GetTemplate :one
+SELECT * FROM templates WHERE id = $1;